requirements.txt
rioxarray=~0.19.0
matplotlib=~3.10.1
icechunk=~0.2.15
jupyterlab
zarr=~3.0.7
xarray=~2025.3.1
leafmap=~0.43.6
localtileserver=~0.10.6
hvplot=~0.11.2
geoviews=~1.14.0
datashader=~0.18.0
jupyter_bokeh=~4.0.5
dask=~2025.4.0
geopandas=~1.0.1
odc-geo=~0.4.10